update to PHP >= 8.2 and phpunit 11.

+++ src/Coordinate/Coordinate.php
@@ -94,7 +94,7 @@ class Coordinate implements CoordinateInterface, \JsonSerializable
      */
     public function normalizeLongitude($longitude)
     {
-        if (180 === floor($longitude) % 360) {
+        if (180 === (int) floor($longitude) % 360) {
             return '180.0';
         }
